requirements.txt
# Sovereign-Doc - Zero-Cost Privacy-First Document Intelligence System
# Core dependencies (required)
pydantic>=2.0.0
loguru>=0.7.0
PyYAML>=6.0
psutil>=5.9.0  # Hardware detection and system monitoring
gputil>=1.4.0  # GPU monitoring (optional - gracefully degrades if no GPU)

# Core Framework
pydantic-settings>=2.0.0

# Async Support
asyncio>=3.4.3

# HTTP Client
httpx>=0.25.0

# Document Processing
opencv-python-headless>=4.8.0
pdf2image>=1.16.0
pypdf>=3.17.0
Pillow>=10.0.0
numpy>=1.24.0

# Computer Vision / Layout Analysis
ultralytics>=8.0.0  # YOLOv8 for document layout detection

# OCR (Optical Character Recognition)
paddleocr>=2.7.0  # Text and table extraction
paddlepaddle>=2.5.0  # PaddleOCR backend (CPU version)

# Transformer Models (for TrOCR handwriting fallback)
transformers>=4.35.0  # TrOCR model
torch>=2.0.0  # PyTorch backend (CPU version)

# Vector Database and Embeddings
# Note: fastembed requires Python <3.13 due to onnxruntime dependency
# For Python 3.14+, these should be installed in a compatible environment
qdrant-client>=1.7.0
# fastembed>=0.2.0  # Commented out due to Python 3.14 incompatibility

# Secure Tunneling
pyngrok>=7.0.0  # ngrok tunnel management for cloud-local communication

# FastAPI (for Cloud Brain server testing)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart
httpx  # Required for FastAPI TestClient

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# UI
streamlit>=1.30.0
